Refactored AutoCreateTable
