+++ sink-connector/src/test/java/com/altinity/clickhouse/sink/connector/db/operations/ClickHouseAutoCreateTableTest.java
@@ -1,137 +1,28 @@
 package com.altinity.clickhouse.sink.connector.db.operations;
 
 import com.altinity.clickhouse.sink.connector.ClickHouseSinkConnectorConfig;
-import com.altinity.clickhouse.sink.connector.db.BaseDbWriter;
-import com.altinity.clickhouse.sink.connector.db.DbWriter;
-import com.altinity.clickhouse.sink.connector.db.HikariDbSource;
 import com.clickhouse.data.ClickHouseDataType;
-import com.clickhouse.jdbc.ClickHouseConnection;
-import io.debezium.data.Json;
-import org.apache.kafka.connect.data.Decimal;
-import org.apache.kafka.connect.data.Field;
-import org.apache.kafka.connect.data.Schema;
-import org.apache.kafka.connect.data.SchemaBuilder;
-import org.junit.After;
+
 import org.junit.Assert;
 import org.junit.jupiter.api.*;
-import org.testcontainers.containers.ClickHouseContainer;
-import org.testcontainers.containers.wait.strategy.HttpWaitStrategy;
-import org.testcontainers.containers.wait.strategy.WaitStrategy;
-import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
-import java.sql.Connection;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
 @Testcontainers
+public class ClickHouseAutoCreateTableTest extends com.altinity.clickhouse.sink.connector.db.operations.ClickHouseAutoCreateTableBase {
 
-public class ClickHouseAutoCreateTableTest {
+    private Map<String, String> columnToDataTypesMap;
+    private ClickHouseSinkConnectorConfig config;
 
-    static Map<String, String> columnToDataTypesMap;
-
-    static Connection conn;
-    
-    ClickHouseSinkConnectorConfig config;
-
-    @Container
-    private ClickHouseContainer clickHouseContainer = new ClickHouseContainer("clickhouse/clickhouse-server:24.8.8")
-            .withInitScript("./init_clickhouse.sql").waitingFor(new HttpWaitStrategy().forPort(8123));
-
-    @AfterAll
-    public static void tearDown() throws SQLException {
-        HikariDbSource.close();
-    }
-    
     @BeforeEach
     public void setUp() {
+        columnToDataTypesMap = getExpectedColumnToDataTypesMap();
         config = new ClickHouseSinkConnectorConfig(new HashMap<>());
     }
-    @BeforeAll
-    static void initialize() {
 
-        columnToDataTypesMap =  getExpectedColumnToDataTypesMap();
-
-//        this.columnToDataTypesMap.put("customer_id", "Int32");
-//        this.columnToDataTypesMap.put("address", "String");
-//        this.columnToDataTypesMap.put("first_name", "String");
-//        this.columnToDataTypesMap.put("amount", "Int32");
-
-        String hostName = "localhost";
-        Integer port = 8123;
-        String database = "test";
-        String userName = "root";
-        String password = "root";
-        String tableName = "auto_create_table";
-
-        ClickHouseSinkConnectorConfig config= new ClickHouseSinkConnectorConfig(new HashMap<>());
-
-
-        String jdbcUrl = BaseDbWriter.getConnectionString(hostName, port, database);
-        Connection conn = DbWriter.createConnection(jdbcUrl, BaseDbWriter.DATABASE_CLIENT_NAME, userName, password,
-                BaseDbWriter.SYSTEM_DB, config);
-        DbWriter writer = new DbWriter(hostName, port, database, tableName, userName, password, config, null, conn);
-
-        conn = writer.getConnection();
-
-    }
-
-    protected Field[] createFields() {
-        ArrayList<Field> fields = new ArrayList<>();
-        fields.add(new Field("customerName", 0, Schema.STRING_SCHEMA));
-        fields.add(new Field("occupation", 1, Schema.STRING_SCHEMA));
-        fields.add(new Field("quantity", 2, Schema.INT32_SCHEMA));
-        fields.add(new Field("amount_1", 3, Schema.FLOAT32_SCHEMA));
-        fields.add(new Field("amount", 4, Schema.FLOAT64_SCHEMA));
-        fields.add(new Field("employed", 5, Schema.BOOLEAN_SCHEMA));
-        fields.add(new Field("blob_storage", 6, SchemaBuilder.type(Schema.BYTES_SCHEMA.type()).
-                name(Decimal.LOGICAL_NAME).build()));
-
-        Schema decimalSchema = SchemaBuilder.type(Schema.BYTES_SCHEMA.type()).parameter("scale", "10")
-                        .parameter("connect.decimal.precision", "30")
-                                .name(Decimal.LOGICAL_NAME).build();
-
-        fields.add(new Field("blob_storage_scale", 7, decimalSchema));
-        fields.add(new Field("json_output", 8, Json.schema()));
-        fields.add(new Field("max_amount", 9, Schema.FLOAT64_SCHEMA));
-
-        Field[] result = new Field[fields.size()];
-        fields.toArray(result);
-        return result;
-    }
-
-    protected static Map<String, String> getExpectedColumnToDataTypesMap() {
-
-        Map<String, String> columnToDataTypesMap = new HashMap<>();
-        columnToDataTypesMap.put("customerName", ClickHouseDataType.String.name());
-        columnToDataTypesMap.put("occupation", ClickHouseDataType.String.name());
-        columnToDataTypesMap.put("quantity", ClickHouseDataType.Int32.name());
-        columnToDataTypesMap.put("amount_1", ClickHouseDataType.Float32.name());
-        columnToDataTypesMap.put("amount", ClickHouseDataType.Float64.name());
-        columnToDataTypesMap.put("employed", ClickHouseDataType.Bool.name());
-        columnToDataTypesMap.put("blob_storage", ClickHouseDataType.String.name());
-        columnToDataTypesMap.put("blob_storage_scale", ClickHouseDataType.Decimal.name());
-        columnToDataTypesMap.put("json_output", ClickHouseDataType.JSON.name());
-
-        columnToDataTypesMap.put("max_amount", ClickHouseDataType.Float64.name());
-
-
-        return columnToDataTypesMap;
-    }
-
-    @Test
-    public void getColumnNameToCHDataTypeMappingTest() {
-        ClickHouseAutoCreateTable act = new ClickHouseAutoCreateTable();
-        Field[] fields = createFields();
-        Map<String, String> colNameToDataTypeMap = act.getColumnNameToCHDataTypeMapping(fields,new ClickHouseSinkConnectorConfig(new HashMap<>()));
-
-        Map<String, String> expectedColNameToDataTypeMap = getExpectedColumnToDataTypesMap();
-
-       // Assert.assertTrue(colNameToDataTypeMap.equals(expectedColNameToDataTypeMap));
-        Assert.assertFalse(colNameToDataTypeMap.isEmpty());
-    }
 
     @Test
     public void testCreateTableSyntax() {
@@ -157,7 +48,8 @@ public class ClickHouseAutoCreateTableTest {
         String query = act.createHistoryTableSyntax(primaryKeys, "auto_create_table_history", "employees",
                 createFields(), this.columnToDataTypesMap, this.config);
         System.out.println("QUERY" + query);
-        Assert.assertTrue(query.equalsIgnoreCase("CREATE TABLE employees.`auto_create_table_history`(`customerName` String NOT NULL,`occupation` String NOT NULL,`quantity` Int32 NOT NULL,`amount_1` Float32 NOT NULL,`amount` Float64 NOT NULL,`employed` Bool NOT NULL,`blob_storage` String NOT NULL,`blob_storage_scale` Decimal NOT NULL,`json_output` JSON NOT NULL,`max_amount` Float64 NOT NULL,`database` String,`table` String,`_raw` String,`_time` UInt64,`is_deleted` UInt8,`operation` String,`_version` UInt64,`host` String,`logfile` String,`position` UInt64,`primary_host` String) ENGINE = MergeTree() ORDER BY(customerName"));
+        String expectedQuery = "CREATE TABLE employees.`auto_create_table_history`(`database` String,`table` String,`table` String,`table` String,`before` String,`after` String,`_raw` String,`_time` UInt64,`is_deleted` UInt8,`operation` String,`_version` UInt64,`host` String,`logfile` String,`position` UInt64,`primary_host` String) ENGINE = MergeTree() ORDER BY(customerName)";
+        Assert.assertTrue(query.equalsIgnoreCase(expectedQuery));
     }
 
     @Test
@@ -187,67 +79,6 @@ public class ClickHouseAutoCreateTableTest {
         System.out.println(query);
     }
 
-    @Test
-    @Tag("IntegrationTest")
-    @Disabled
-    public void testCreateNewTable() {
-        String dbHostName = clickHouseContainer.getHost();
-        Integer port = clickHouseContainer.getFirstMappedPort();
-        String database = "test";
-        String userName = clickHouseContainer.getUsername();
-        String password = clickHouseContainer.getPassword();
-        String tableName = "employees";
-
-
-        String jdbcUrl = BaseDbWriter.getConnectionString(dbHostName, port, database);
-        Connection conn = DbWriter.createConnection(jdbcUrl, BaseDbWriter.DATABASE_CLIENT_NAME, userName, password,
-                BaseDbWriter.SYSTEM_DB, new ClickHouseSinkConnectorConfig(new HashMap<>()));
-
-
-
-        DbWriter writer = new DbWriter(dbHostName, port, database, tableName, userName, password,
-                new ClickHouseSinkConnectorConfig(new HashMap<>()), null, conn);
-
-        ClickHouseAutoCreateTable act = new ClickHouseAutoCreateTable();
-        ArrayList<String> primaryKeys = new ArrayList<>();
-        primaryKeys.add("customerName");
-
-        try {
-            act.createNewTable(primaryKeys, tableName, "test", this.createFields(), writer.getConnection(),
-                    false, false, null,new ClickHouseSinkConnectorConfig(new HashMap<>()));
-        } catch(SQLException se) {
-            Assert.assertTrue(false);
-        }
-    }
-
-    @Test
-    @Tag("IntegrationTest")
-    @Disabled
-    public void testCreateMergeTreeHistoryTable() {
-        String dbHostName = clickHouseContainer.getHost();
-        Integer port = clickHouseContainer.getFirstMappedPort();
-        String database = "test";
-        String userName = clickHouseContainer.getUsername();
-        String password = clickHouseContainer.getPassword();
-        String tableName = "employees5_history";
-
-        String jdbcUrl = BaseDbWriter.getConnectionString(dbHostName, port, database);
-        Connection conn = DbWriter.createConnection(jdbcUrl, BaseDbWriter.DATABASE_CLIENT_NAME, userName, password,
-                BaseDbWriter.SYSTEM_DB, new ClickHouseSinkConnectorConfig(new HashMap<>()));
-
-        DbWriter writer = new DbWriter(dbHostName, port, database, tableName, userName, password,
-                new ClickHouseSinkConnectorConfig(new HashMap<>()), null, conn);
-
-        ClickHouseAutoCreateTable act = new ClickHouseAutoCreateTable();
-        ArrayList<String> primaryKeys = new ArrayList<>();
-        primaryKeys.add("customerName");
-
-        try {
-            act.createHistoryTable(primaryKeys, tableName, "test", this.createFields(), writer.getConnection(), this.config);
-        } catch(SQLException se) {
-            Assert.assertTrue(false);
-        }
-    }
 
     @Test
     public void testIsPrimaryKeyColumnPresent()    {
